test(terminal): split 1154-line test_terminal_service.py by lifecycle phase (#109, A-8)
Splits services/terminal/tests/unit/test_terminal_service.py (1154
lines, 16 test classes + 5 module-level helpers) into three focused
files. Each split includes all the original helpers
(make_terminal / make_service / _setup_transaction_mock /
_make_idle_terminal / _make_opened_terminal) so they can be run in
isolation.

  test_terminal_service_lifecycle.py (579 lines) — state-mutating ops:
    Create, Delete, UpdateDescription, UpdateFunctionMode, SignInOut,
    OpenTerminal, CloseTerminal.

  test_terminal_service_query.py     (329 lines) — read-only ops:
    GetAndStatus, List, GetInfo, ConvertDatetime, GetStoreName.

  test_terminal_service_logging.py   (533 lines) — pubsub-side:
    CashInOut, Publish, DeliveryStatus, RepublishUndelivered.

Verified: pipenv run pytest -m unit → 332 passed (same as before).

Refs #109, follows up #113 (other large files: master/test_master_services.py,
cart/test_cart.py, etc. — same pattern when needed).
